Make tin-summer compatible with Rust stable
To do so we:
  - move benchmarks to `/benchmarks`, so that they are not build during
    `cargo install` (this might help with binary size as well perhaps?)
  - replace `AtomicU64` with `AtomicUsize`, which is more or less the
    same thing (there's no AtomicU64 on x86_32)!
